+++ MS5611.h
@@ -62,9 +62,10 @@ public:
   bool     begin(TwoWire *wire = &Wire);
   bool     isConnected();
 
-  //  reset command + get constants
-  //  returns false if ROM constants == 0;
-  bool     reset();
+  //       reset command + get constants
+  //       mathMode = 0 (default), 1 = factor 2 fix.
+  //       returns false if ROM constants are 0;
+  bool     reset(uint8_t mathMode = 0);
 
   //  the actual reading of the sensor;
   //  returns MS5611_READ_OK upon success
@@ -113,17 +114,13 @@ public:
   uint16_t getManufacturer();
   uint16_t getSerialCode();
 
-  //       only use 
-  //       0 = default, 1 = factor 2 fix.
-  void     setMathMode(uint8_t mathMode);
-  uint8_t  getMathMode();
 
 protected:
   void     convert(const uint8_t addr, uint8_t bits);
   uint32_t readADC();
   uint16_t readProm(uint8_t reg);
   int      command(const uint8_t command);
-  void     initConstants();
+  void     initConstants(uint8_t mathMode);
 
   uint8_t  _address;
   uint8_t  _samplingRate;
@@ -136,7 +133,6 @@ protected:
   uint32_t _lastRead;
   uint32_t _deviceID;
   bool     _compensation;
-  uint8_t  _mathMode = 0;
 
   TwoWire * _wire;
 };